Separate out air con control from auxiliaries.*
