+++ speeduino/src/controllers/aircon/airconController.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void initialiseAirCon(void);
+
+void airConControl(void);
