+++ test/test_accontrol/shared.h
@@ -10,9 +10,9 @@ extern uint8_t acTPSLockoutDelay;
 extern uint8_t acRPMLockoutDelay;
 extern uint8_t acAfterEngineStartDelay;
 extern bool waitedAfterCranking;
-extern boardOutputPin_t aircon_comp_pin;
-extern boardOutputPin_t aircon_fan_pin;
-extern fastInputPin_t aircon_req_pin;
+extern outputPin_t aircon_comp_pin;
+extern outputPin_t aircon_fan_pin;
+extern inputPin_t aircon_req_pin;
 
 constexpr uint8_t TEST_ACREQUEST_PIN = 11;
 constexpr uint8_t TEST_ACCOMP_PIN = 12;
