+++ test/test_accontrol/test_airconcontrol.cpp
@@ -1,6 +1,6 @@
 #include "../test_utils.h"
 #include "globals.h"
-#include "auxiliaries.h"
+#include "src/controllers/aircon/airconController.h"
 #include "shared.h"
 #include "units.h"
